+++ tests/test_views.py
@@ -16,12 +16,19 @@ from django.http import HttpRequest
 from django.test import TestCase, override_settings
 from django.test.client import Client
 from django.urls import reverse
+from freezegun import freeze_time
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey
 
 from learning_assistant.models import LearningAssistantAuditTrial, LearningAssistantMessage
 
 User = get_user_model()
+
+mocked_course = {
+    'course': 'edx+test',
+    'start': None,
+    'end': None,
+}
 
 
 class FakeClient(Client):
@@ -76,8 +83,23 @@ class LoggedInTestCase(TestCase):
         self.user.save()
         self.client.login_user(self.user)
 
+        self.mocked_get_cache_course_run_data = patch(
+            'learning_assistant.views.get_cache_course_run_data',
+            return_value=mocked_course
+        )
+        self.mocked_get_cache_course_run_data.start()
+        self.addCleanup(self.mocked_get_cache_course_run_data.stop)
+
+        self.mocked_get_course_id = patch(
+            'learning_assistant.views.get_course_id',
+            return_value=mocked_course['course']
+        )
+        self.mocked_get_course_id.start()
+        self.addCleanup(self.mocked_get_course_id.stop)
+
 
 @ddt.ddt
+@freeze_time('2024-06-15')
 class CourseChatViewTests(LoggedInTestCase):
     """
     Test for the CourseChatView
@@ -91,12 +113,6 @@ class CourseChatViewTests(LoggedInTestCase):
         super().setUp()
         self.course_id = 'course-v1:edx+test+23'
         self.course_run_key = CourseKey.from_string(self.course_id)
-
-        self.patcher = patch(
-            'learning_assistant.api.get_cache_course_run_data',
-            return_value={'course': 'edx+test'}
-        )
-        self.patcher.start()
 
     @patch('learning_assistant.views.learning_assistant_enabled')
     def test_course_waffle_inactive(self, mock_waffle):
@@ -129,6 +145,87 @@ class CourseChatViewTests(LoggedInTestCase):
             content_type='application/json'
         )
         self.assertEqual(response.status_code, 400)
+
+    @ddt.data({
+        'start': '2023-01-01T01:00:00Z',  # past date
+        'end': '2023-01-30T01:00:00Z',    # past date
+        'expects_fail': True,
+    }, {
+        'start': '2026-01-01T01:00:00Z',  # future date
+        'end': '2026-01-30T01:00:00Z',    # future date
+        'expects_fail': True,
+    }, {
+        'start': '2024-06-01T01:00:00Z',  # past date
+        'end': '2024-06-01T30:00:00Z',    # future date
+        'expects_fail': False,
+    }, {
+        'start': '2024-06-01T01:00:00Z',  # past date
+        'end': None,
+        'expects_fail': False,
+    }, {
+        'start': None,
+        'end': '2024-06-01T30:00:00Z',    # future date
+        'expects_fail': False,
+    })
+    @patch('learning_assistant.views.get_cache_course_run_data')
+    @patch('learning_assistant.views.audit_trial_is_expired')
+    @patch('learning_assistant.views.render_prompt_template')
+    @patch('learning_assistant.views.get_chat_response')
+    @patch('learning_assistant.views.learning_assistant_enabled')
+    @patch('learning_assistant.views.get_user_role')
+    @patch('learning_assistant.views.CourseEnrollment')
+    @patch('learning_assistant.views.CourseMode')
+    @patch('learning_assistant.views.chat_history_enabled')
+    @override_settings(LEARNING_ASSISTANT_PROMPT_TEMPLATE='This is the default template')
+    def test_dates_for_chat_response(
+        self,
+        dates,
+        mock_chat_history_enabled,
+        mock_mode,
+        mock_enrollment,
+        mock_get_user_role,
+        mock_waffle,
+        mock_chat_response,
+        mock_render,
+        mock_trial_expired,
+        mocked_get_cache_course_run_data,
+    ):
+        mocked_get_cache_course_run_data.return_value = {
+            'course': mocked_course['course'],
+            'start': dates['start'],
+            'end': dates['end'],
+        }
+        mock_waffle.return_value = True
+        mock_get_user_role.return_value = 'staff'
+        mock_mode.VERIFIED_MODES = ['verified']
+        mock_mode.CREDIT_MODES = ['credit']
+        mock_mode.NO_ID_PROFESSIONAL_MODE = 'no-id'
+        mock_mode.UPSELL_TO_VERIFIED_MODES = ['audit']
+        mock_enrollment.get_enrollment.return_value = MagicMock(mode='staff')
+        mock_chat_response.return_value = (200, {'role': 'assistant', 'content': 'Something else'})
+        mock_render.return_value = 'Rendered template mock'
+        mock_trial_expired.return_value = False
+        test_unit_id = 'test-unit-id'
+
+        mock_chat_history_enabled.return_value = True
+
+        test_data = [
+            {'role': 'user', 'content': 'What is 2+2?'},
+            {'role': 'assistant', 'content': 'It is 4'},
+            {'role': 'user', 'content': 'And what else?'},
+        ]
+
+        response = self.client.post(
+            reverse('chat', kwargs={'course_run_id': self.course_id})+f'?unit_id={test_unit_id}',
+            data=json.dumps(test_data),
+            content_type='application/json'
+        )
+
+        if dates['expects_fail']:
+            self.assertEqual(response.status_code, 403)
+            self.assertEqual(response.data, {'detail': 'Learning assistant not enabled for course.'})
+        else:
+            self.assertEqual(response.status_code, 200)
 
     @patch('learning_assistant.views.audit_trial_is_expired')
     @patch('learning_assistant.views.learning_assistant_enabled')
@@ -249,222 +346,7 @@ class CourseChatViewTests(LoggedInTestCase):
 
 
 @ddt.ddt
-class LearningAssistantEnabledViewTests(LoggedInTestCase):
-    """
-    Test for the LearningAssistantEnabledView
-    """
-    sys.modules['lms.djangoapps.courseware.access'] = MagicMock()
-    sys.modules['lms.djangoapps.courseware.toggles'] = MagicMock()
-    sys.modules['common.djangoapps.course_modes.models'] = MagicMock()
-    sys.modules['common.djangoapps.student.models'] = MagicMock()
-
-    def setUp(self):
-        super().setUp()
-        self.course_id = 'course-v1:edx+test+23'
-
-    @ddt.data(
-        (True, True),
-        (False, False),
-    )
-    @ddt.unpack
-    @patch('learning_assistant.views.learning_assistant_enabled')
-    def test_learning_assistant_enabled(self, mock_value, message, mock_learning_assistant_enabled):
-        mock_learning_assistant_enabled.return_value = mock_value
-        response = self.client.get(reverse('enabled', kwargs={'course_run_id': self.course_id}))
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {'enabled': message})
-
-    @patch('learning_assistant.views.learning_assistant_enabled')
-    def test_invalid_course_id(self, mock_learning_assistant_enabled):
-        mock_learning_assistant_enabled.return_value = True
-        response = self.client.get(reverse('enabled', kwargs={'course_run_id': self.course_id+'+invalid'}))
-
-        self.assertEqual(response.status_code, 400)
-
-
-class LearningAssistantMessageHistoryViewTests(LoggedInTestCase):
-    """
-    Tests for the LearningAssistantMessageHistoryView
-    """
-
-    def setUp(self):
-        super().setUp()
-        self.course_id = 'course-v1:edx+test+23'
-
-    @patch('learning_assistant.views.learning_assistant_enabled')
-    def test_course_waffle_inactive(self, mock_waffle):
-        mock_waffle.return_value = False
-        message_count = 5
-        response = self.client.get(
-            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={message_count}',
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 403)
-
-    @patch('learning_assistant.views.learning_assistant_enabled')
-    def test_learning_assistant_not_enabled(self, mock_learning_assistant_enabled):
-        mock_learning_assistant_enabled.return_value = False
-        message_count = 5
-        response = self.client.get(
-            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={message_count}',
-            content_type='application/json'
-        )
-
-        self.assertEqual(response.status_code, 403)
-
-    @patch('learning_assistant.views.chat_history_enabled')
-    @patch('learning_assistant.views.learning_assistant_enabled')
-    @patch('learning_assistant.views.get_user_role')
-    @patch('learning_assistant.views.CourseEnrollment')
-    @patch('learning_assistant.views.CourseMode')
-    def test_user_no_enrollment_not_staff(
-        self,
-        mock_mode,
-        mock_enrollment,
-        mock_get_user_role,
-        mock_assistant_waffle,
-        mock_history_waffle
-    ):
-        mock_assistant_waffle.return_value = True
-        mock_history_waffle.return_value = True
-        mock_get_user_role.return_value = 'student'
-        mock_mode.VERIFIED_MODES = ['verified']
-        mock_enrollment.get_enrollment = MagicMock(return_value=None)
-
-        message_count = 5
-        response = self.client.get(
-            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={message_count}',
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 403)
-
-    @patch('learning_assistant.views.chat_history_enabled')
-    @patch('learning_assistant.views.learning_assistant_enabled')
-    @patch('learning_assistant.views.get_user_role')
-    @patch('learning_assistant.views.CourseEnrollment')
-    @patch('learning_assistant.views.CourseMode')
-    def test_user_audit_enrollment_not_staff(
-        self,
-        mock_mode,
-        mock_enrollment,
-        mock_get_user_role,
-        mock_assistant_waffle,
-        mock_history_waffle
-    ):
-        mock_assistant_waffle.return_value = True
-        mock_history_waffle.return_value = True
-        mock_get_user_role.return_value = 'student'
-        mock_mode.VERIFIED_MODES = ['verified']
-        mock_enrollment.get_enrollment.return_value = MagicMock(mode='audit')
-
-        message_count = 5
-        response = self.client.get(
-            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={message_count}',
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 403)
-
-    @patch('learning_assistant.views.chat_history_enabled')
-    @patch('learning_assistant.views.learning_assistant_enabled')
-    @patch('learning_assistant.views.get_user_role')
-    @patch('learning_assistant.views.CourseEnrollment')
-    @patch('learning_assistant.views.CourseMode')
-    @patch('learning_assistant.views.get_course_id')
-    def test_learning_message_history_view_get(
-        self,
-        mock_get_course_id,
-        mock_mode,
-        mock_enrollment,
-        mock_get_user_role,
-        mock_assistant_waffle,
-        mock_history_waffle,
-    ):
-        mock_assistant_waffle.return_value = True
-        mock_history_waffle.return_value = True
-        mock_get_user_role.return_value = 'student'
-        mock_mode.VERIFIED_MODES = ['verified']
-        mock_enrollment.get_enrollment.return_value = MagicMock(mode='verified')
-
-        LearningAssistantMessage.objects.create(
-            course_id=self.course_id,
-            user=self.user,
-            role='staff',
-            content='Older message',
-            created=date(2024, 10, 1)
-        )
-
-        LearningAssistantMessage.objects.create(
-            course_id=self.course_id,
-            user=self.user,
-            role='staff',
-            content='Newer message',
-            created=date(2024, 10, 3)
-        )
-
-        db_messages = LearningAssistantMessage.objects.all().order_by('created')
-        db_messages_count = len(db_messages)
-
-        mock_get_course_id.return_value = self.course_id
-        response = self.client.get(
-            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={db_messages_count}',
-            content_type='application/json'
-        )
-        data = response.data
-
-        # Ensure same number of entries
-        self.assertEqual(len(data), db_messages_count)
-
-        # Ensure values are as expected
-        for i, message in enumerate(data):
-            self.assertEqual(message['role'], db_messages[i].role)
-            self.assertEqual(message['content'], db_messages[i].content)
-            self.assertEqual(message['timestamp'], db_messages[i].created.isoformat())
-
-    @patch('learning_assistant.views.chat_history_enabled')
-    @patch('learning_assistant.views.learning_assistant_enabled')
-    @patch('learning_assistant.views.get_course_id')
-    def test_learning_message_history_view_get_disabled(
-        self,
-        mock_get_course_id,
-        mock_assistant_waffle,
-        mock_history_waffle,
-    ):
-        mock_assistant_waffle.return_value = True
-        mock_history_waffle.return_value = False
-
-        LearningAssistantMessage.objects.create(
-            course_id=self.course_id,
-            user=self.user,
-            role='staff',
-            content='Older message',
-            created=date(2024, 10, 1)
-        )
-
-        LearningAssistantMessage.objects.create(
-            course_id=self.course_id,
-            user=self.user,
-            role='staff',
-            content='Newer message',
-            created=date(2024, 10, 3)
-        )
-
-        db_messages = LearningAssistantMessage.objects.all().order_by('created')
-        db_messages_count = len(db_messages)
-
-        mock_get_course_id.return_value = self.course_id
-        response = self.client.get(
-            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={db_messages_count}',
-            content_type='application/json'
-        )
-        data = response.data
-
-        # Ensure returning an empty list
-        self.assertEqual(len(data), 0)
-        self.assertEqual(data, [])
-
-
-@ddt.ddt
+@freeze_time('2024-06-15')
 class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
     """
     Tests for the LearningAssistantChatSummaryView
@@ -487,6 +369,7 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response.data['detail'], 'Course ID is not a valid course ID.')
 
+    @freeze_time('2024-01-01')
     @ddt.data(
         *product(
             [True, False],                                   # learning assistant enabled
@@ -561,11 +444,13 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
         mock_audit_trial_is_expired.return_value = audit_trial_is_expired_mock_value
         mock_get_audit_trial_length_days.return_value = audit_trial_length_days_mock_value
 
-        trial_start_date = datetime(2024, 1, 1, 0, 0, 0)
+        audit_trial_start_date = datetime.now()
+        audit_trial_expiration_date = audit_trial_start_date + timedelta(days=audit_trial_length_days_mock_value)
         if trial_available:
             LearningAssistantAuditTrial.objects.create(
                 user=self.user,
-                start_date=trial_start_date,
+                start_date=audit_trial_start_date,
+                expiration_date=audit_trial_expiration_date,
             )
 
         url_kwargs = {'course_run_id': self.course_id}
@@ -577,8 +462,17 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
 
         response = self.client.get(url)
 
-        # Assert message history data is correct.
+        if not learning_assistant_enabled_mock_value:
+            self.assertEqual(response.data, {
+                'enabled': False,
+                'message_history': [],
+                'audit_trial': {},
+                'audit_trial_length_days': 0,
+            })
+            return
+
         if chat_history_enabled_mock_value:
+            # Assert message history data is correct.
             data = response.data['message_history']
 
             # Ensure same number of entries.
@@ -595,12 +489,13 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
         # Assert trial data is correct.
         expected_trial_data = {}
         if trial_available:
-            expected_trial_data['start_date'] = trial_start_date
-            expected_trial_data['expiration_date'] = trial_start_date + timedelta(days=14)
+            expected_trial_data['start_date'] = audit_trial_start_date
+            expected_trial_data['expiration_date'] = audit_trial_expiration_date
 
         self.assertEqual(response.data['audit_trial'], expected_trial_data)
         self.assertEqual(response.data['audit_trial_length_days'], audit_trial_length_days_mock_value)
 
+    @freeze_time('2024-01-01')
     @ddt.data(
         *product(
             [True, False],                    # learning assistant enabled
@@ -675,11 +570,13 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
         mock_audit_trial_is_expired.return_value = audit_trial_is_expired_mock_value
         mock_get_audit_trial_length_days.return_value = audit_trial_length_days_mock_value
 
-        trial_start_date = datetime(2024, 1, 1, 0, 0, 0)
+        audit_trial_start_date = datetime.now()
+        audit_trial_expiration_date = audit_trial_start_date + timedelta(days=audit_trial_length_days_mock_value)
         if trial_available:
             LearningAssistantAuditTrial.objects.create(
                 user=self.user,
-                start_date=trial_start_date,
+                start_date=audit_trial_start_date,
+                expiration_date=audit_trial_expiration_date,
             )
 
         url_kwargs = {'course_run_id': self.course_id}
@@ -691,8 +588,17 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
 
         response = self.client.get(url)
 
-        # Assert message history data is correct.
+        if not learning_assistant_enabled_mock_value:
+            self.assertEqual(response.data, {
+                'enabled': False,
+                'message_history': [],
+                'audit_trial': {},
+                'audit_trial_length_days': 0,
+            })
+            return
+
         if chat_history_enabled_mock_value:
+            # Assert message history data is correct.
             data = response.data['message_history']
 
             # Ensure same number of entries.
@@ -709,12 +615,13 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
         # Assert trial data is correct.
         expected_trial_data = {}
         if trial_available:
-            expected_trial_data['start_date'] = trial_start_date
-            expected_trial_data['expiration_date'] = trial_start_date + timedelta(days=14)
+            expected_trial_data['start_date'] = audit_trial_start_date
+            expected_trial_data['expiration_date'] = audit_trial_expiration_date
 
         self.assertEqual(response.data['audit_trial'], expected_trial_data)
         self.assertEqual(response.data['audit_trial_length_days'], audit_trial_length_days_mock_value)
 
+    @freeze_time('2024-01-01')
     @ddt.data(
         *product(
             [True, False],  # learning assistant enabled
@@ -789,11 +696,13 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
         mock_audit_trial_is_expired.return_value = audit_trial_is_expired_mock_value
         mock_get_audit_trial_length_days.return_value = audit_trial_length_days_mock_value
 
-        trial_start_date = datetime(2024, 1, 1, 0, 0, 0)
+        audit_trial_start_date = datetime.now()
+        audit_trial_expiration_date = audit_trial_start_date + timedelta(days=audit_trial_length_days_mock_value)
         if trial_available:
             LearningAssistantAuditTrial.objects.create(
                 user=self.user,
-                start_date=trial_start_date,
+                start_date=audit_trial_start_date,
+                expiration_date=audit_trial_expiration_date,
             )
 
         url_kwargs = {'course_run_id': self.course_id}
@@ -804,6 +713,15 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
             url = f"{url}?{urlencode(query_params)}"
 
         response = self.client.get(url)
+
+        if not learning_assistant_enabled_mock_value:
+            self.assertEqual(response.data, {
+                'enabled': False,
+                'message_history': [],
+                'audit_trial': {},
+                'audit_trial_length_days': 0,
+            })
+            return
 
         # Assert message history data is correct.
         if chat_history_enabled_mock_value and trial_available and not audit_trial_is_expired_mock_value:
@@ -823,8 +741,68 @@ class LearningAssistantChatSummaryViewTests(LoggedInTestCase):
         # Assert trial data is correct.
         expected_trial_data = {}
         if trial_available:
-            expected_trial_data['start_date'] = trial_start_date
-            expected_trial_data['expiration_date'] = trial_start_date + timedelta(days=14)
+            expected_trial_data['start_date'] = audit_trial_start_date
+            expected_trial_data['expiration_date'] = audit_trial_expiration_date
 
         self.assertEqual(response.data['audit_trial'], expected_trial_data)
         self.assertEqual(response.data['audit_trial_length_days'], audit_trial_length_days_mock_value)
+
+    @ddt.data({
+        'start': '2023-01-01T01:00:00Z',  # past date
+        'end': '2023-01-30T01:00:00Z',    # past date
+        'expects_fail': True,
+    }, {
+        'start': '2026-01-01T01:00:00Z',  # future date
+        'end': '2026-01-30T01:00:00Z',    # future date
+        'expects_fail': True,
+    }, {
+        'start': '2024-06-01T01:00:00Z',  # past date
+        'end': '2024-06-01T30:00:00Z',    # future date
+        'expects_fail': False,
+    }, {
+        'start': '2024-06-01T01:00:00Z',  # past date
+        'end': None,
+        'expects_fail': False,
+    }, {
+        'start': None,
+        'end': '2024-06-01T30:00:00Z',    # future date
+        'expects_fail': False,
+    })
+    @patch('learning_assistant.views.get_cache_course_run_data')
+    @patch('learning_assistant.views.chat_history_enabled')
+    @patch('learning_assistant.views.learning_assistant_enabled')
+    @patch('learning_assistant.views.get_user_role')
+    @patch('learning_assistant.views.CourseEnrollment')
+    @patch('learning_assistant.views.CourseMode')
+    def test_chat_summary_with_date_gating(
+        self,
+        dates,
+        mock_mode,
+        mock_enrollment,
+        mock_get_user_role,
+        mock_learning_assistant_enabled,
+        mock_chat_history_enabled,
+        mocked_get_cache_course_run_data,
+    ):
+        mocked_get_cache_course_run_data.return_value = {
+            'course': mocked_course['course'],
+            'start': dates['start'],
+            'end': dates['end'],
+        }
+        mock_learning_assistant_enabled.return_value = True
+        mock_chat_history_enabled.return_value = True
+        mock_get_user_role.return_value = 'staff'
+        mock_mode.VERIFIED_MODES = ['verified']
+        mock_mode.CREDIT_MODES = ['credit']
+        mock_mode.NO_ID_PROFESSIONAL_MODE = 'no-id'
+        mock_mode.UPSELL_TO_VERIFIED_MODES = ['audit']
+
+        mock_enrollment.get_enrollment.return_value = MagicMock(mode='staff')
+
+        url_kwargs = {'course_run_id': self.course_id}
+        url = reverse('chat-summary', kwargs=url_kwargs)
+
+        response = self.client.get(url)
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.data['enabled'], not dates['expects_fail'])